requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv==0.15.0
asyncio==3.4.3
aiohttp==3.6.3
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
